requirements.txt
numpy
Pillow
screen-brightness-control
sched
time